# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "infinitas hosted registry",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/releases/{release_id}/exposures": {
      "get": {
        "tags": [
          "exposure"
        ],
        "summary": "List Exposures",
        "operationId": "list_exposures_api_v1_releases__release_id__exposures_get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExposureView"
                  },
                  "title": "Response List Exposures Api V1 Releases  Release Id  Exposures Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "exposure"
        ],
        "summary": "Create Exposure",
        "operationId": "create_exposure_api_v1_releases__release_id__exposures_post",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExposureCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/exposures/{exposure_id}": {
      "patch": {
        "tags": [
          "exposure"
        ],
        "summary": "Patch Exposure",
        "operationId": "patch_exposure_api_v1_exposures__exposure_id__patch",
        "parameters": [
          {
            "name": "exposure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exposure Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExposurePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/exposures/{exposure_id}/activate": {
      "post": {
        "tags": [
          "exposure"
        ],
        "summary": "Activate Exposure",
        "operationId": "activate_exposure_api_v1_exposures__exposure_id__activate_post",
        "parameters": [
          {
            "name": "exposure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exposure Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/exposures/{exposure_id}/revoke": {
      "post": {
        "tags": [
          "exposure"
        ],
        "summary": "Revoke Exposure",
        "operationId": "revoke_exposure_api_v1_exposures__exposure_id__revoke_post",
        "parameters": [
          {
            "name": "exposure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exposure Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExposureCreateRequest": {
        "properties": {
          "audience_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "private",
                  "grant",
                  "authenticated",
                  "public"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Type"
          },
          "listing_mode": {
            "type": "string",
            "enum": [
              "listed",
              "direct_only"
            ],
            "title": "Listing Mode",
            "default": "listed"
          },
          "install_mode": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Install Mode",
            "default": "enabled"
          },
          "requested_review_mode": {
            "type": "string",
            "enum": [
              "none",
              "advisory",
              "blocking"
            ],
            "title": "Requested Review Mode",
            "default": "none"
          }
        },
        "type": "object",
        "title": "ExposureCreateRequest"
      },
      "ExposurePatchRequest": {
        "properties": {
          "listing_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "listed",
                  "direct_only"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Mode"
          },
          "install_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "enabled",
                  "disabled"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Mode"
          },
          "requested_review_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "advisory",
                  "blocking"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Review Mode"
          }
        },
        "type": "object",
        "title": "ExposurePatchRequest"
      },
      "ExposureView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "release_id": {
            "type": "integer",
            "title": "Release Id"
          },
          "audience_type": {
            "type": "string",
            "title": "Audience Type"
          },
          "listing_mode": {
            "type": "string",
            "title": "Listing Mode"
          },
          "install_mode": {
            "type": "string",
            "title": "Install Mode"
          },
          "review_requirement": {
            "type": "string",
            "title": "Review Requirement"
          },
          "requested_review_mode": {
            "type": "string",
            "title": "Requested Review Mode"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "requested_by_principal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Principal Id"
          },
          "policy_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy Snapshot"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "release_id",
          "audience_type",
          "listing_mode",
          "install_mode",
          "review_requirement",
          "requested_review_mode",
          "state",
          "policy_snapshot"
        ],
        "title": "ExposureView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "infinitas_auth_token"
      }
    }
  }
}